+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Handlers listening for `apostrophe:run` event are now able to send exit signal to the Apostrophe bootstrap routine.
 * Adds new event `@apostrophecms/doc:afterAllModesDeleted` fired after all modes of a given document are purged.
 * Support for Node.js 17 and 18. MongoDB connections to `localhost` will now successfully find a typical dev MongoDB server bound only to `127.0.0.1`, Apostrophe can generate valid ipv6 URLs pointing back to itself, and `webpack` and `vue-loader` have been updated to address incompatibilities.
+* Adds support for custom context menus provided by any module (see `apos.doc.addContextOperation()`).
 
 ### Fixes